adds and refactors runsConfig test cases
